version 20210224-1

+++ commands/del.js
@@ -2,14 +2,18 @@ module.exports = {
     name: 'del',
     description: 'Delete message',
     execute(message, args) {
-        const amount = parseInt(args[0]) + 1;
-        if (isNaN(amount)) {
-            return message.reply('Ce n\'est pas un nombre valide !');
-        }
-        else if (amount <= 1 || amount > 100) {
-            return message.reply('Ce n\'est pas un nombre valide ! (entre 1 et 100)');
-        }
-        message.channel.bulkDelete(amount)
-            .then(messages => console.log(`${messages.size - 1} messages supprimés`))
+        if (message.member.permissions.has('MANAGE_MESSAGES', 1)) {
+            const amount = parseInt(args[0]) + 1;
+            if (isNaN(amount)) {
+                return message.reply('Ce n\'est pas un nombre valide !');
+            }
+            else if (amount <= 1 || amount > 100) {
+                return message.reply('Ce n\'est pas un nombre valide ! (entre 1 et 99)');
+            }
+            message.channel.bulkDelete(amount)
+                .then(messages => {
+                    console.log(`${messages.size - 1} messages supprimés par ${message.author.username} sur ${message.guild.name} dans ${message.channel.name}`);
+                });
+        } else { message.reply('Tu n\'est pas modérateur !');}
     }
 }
